+++ lib/vnet/reexec_daemon_darwin.go
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-//go:build !darwin && !windows
-// +build !darwin,!windows
+//go:build vnetdaemon
+// +build vnetdaemon
 
 package vnet
 
@@ -23,12 +23,17 @@ import (
 	"context"
 
 	"github.com/gravitational/trace"
+
+	"github.com/gravitational/teleport/lib/vnet/daemon"
 )
 
-func configureOS(ctx context.Context, cfg *osConfig) error {
-	return trace.Wrap(ErrVnetNotImplemented)
+// execAdminProcess is called from the normal user process to execute the
+// register and call the daemon process.
+func execAdminProcess(ctx context.Context, config daemon.Config) error {
+	return trace.Wrap(daemon.RegisterAndCall(ctx, config))
 }
 
-func (c *osConfigurator) doWithDroppedRootPrivileges(ctx context.Context, fn func() error) (err error) {
-	return trace.Wrap(ErrVnetNotImplemented)
+// DaemonSubcommand runs the VNet daemon process.
+func DaemonSubcommand(ctx context.Context) error {
+	return trace.Wrap(daemon.Start(ctx, RunAdminProcess))
 }
